+++ CHANGELOG.rst
@@ -12,7 +12,7 @@ Version 0.4 - 2012-Sep-07
   a dictionary.
 * Socket and WSGI :class:`pulsar.Application` are built on top of the new
   :class:`pulsar.AsyncSocketServer` framework class.
-* **206 regression tests**, **79% coverage**.
+* **208 regression tests**, **79% coverage**.
 
 Version 0.3 - 2012-May-03
 ============================